Remove PEP-257 errors
The script has a number of PEP-257 violations which are being used
to demonstrate/test its functionality.  However this means it's
difficult to deploy the script in an environment where all Python
files, including the script itself, will be automatically checked
for PEP-257 compliance.

The errors are removed and will be added as separate tests.