Fix "First line should be in imperative mood" warnings
When pep257.py is run on itself, warnings about imperative mood
are reported on the docstrings of a few of the test methods.

These are false-positive warnings.  Rather than attempting to
update the check to catch all the edge cases like these, modify
the docstrings slightly to avoid the warnings.

Fixes #25

Change-Id: Ic0084f9bdb881de83ab77e516f0bb9195882c359