Recursively process subfolders also when used as imported module
In commit 5624b71 the command line argument handling was modified
to recursively add files from subfolders passed as arguments.

However the same functionality was not added for the case that
the pep257.py module is imported and the check_files method is
used, for example:

 import pep257
 errors = pep257.check_files(["file1.py", "folder"])

Refactor the previous change so that the recursive subfolder
handling is used in both cases.

Change-Id: I05f772916cdaff5fd7d2a00fcbbe69b1ea0ab79a