Use file's basename when checking if it's a test file
If the script is invoked with a folder name in the arguments,
for example:

 ./pep257.py folder

Then the list of files generated contains filenames prefixed
with the folder name, for example:

 folder/file1.py
 folder/file2.py
 folder/test_file1.py

This causes the check `filename.startswith("_test")` to always
return False, and test files are not properly handled.

Use `os.path.basename` to strip off any leading folder names
before checking if the file is a test file.

Change-Id: I05ccb1c168eb70ab75c8c28506ce198848355097