Exit with error status when PEP257 violations are found
The pep257.py script currently always exits with status code 0 whether
or not there are PEP-257 violations in the input file(s).  This makes
it difficult to use in automated checks that rely on the exit status,
such as running the script in Jenkins jobs triggered from Gerrit.

Update the script to exit with an error code when there are errors.

Change-Id: I4fd7b0c80cebfea5d8daf353dba9d1ee440b041f
Signed-off-by: David Pursehouse <david.pursehouse@sonymobile.com>