Fix tests.
Test worked on first run and failed on the second run.
The reason for this is the reference to `__file__` in `test.py`.
The problem is that after the first run, `__file__` points to the `.pyc` file instead of the original `.py` file which makes `pep257` fail (because it requires the source).