Fix parser test case for invalid source
Commit e886115 introduced some tests that intend to iterate over strings
that contain definitions of invalid source code and then assert that the
parser does something sensible with these sources.
However, there was a typo in e886115 causing the tests to not execute as
expected: the invalid sources are iterated in the `source_ucli`
variable. However, the parsing is done against the `source_ucl` variable
which contains a known parsable source string!