Default value for breaking API change.
Some tools like prospector tried to import a v1 pydocstyle or a v2 pydocstyle and
uses them the same way, BUT there has been a breaking change introduced here and so
`ignore_decorators` must be provided. This breaks at least prospectors behavior. 

May we set just a default=None value to have the same signature for the v1 and v2 function here.