Fix failing tests due to modifications to imperatives
With the recent addition of "initiate" to the list of imperatives,
both "initialize" and "initiate" stem to the same verb - "initi"
which causes an error since we use Initialize in our docs and due
to the way the imperative suggestion logic is implemented.

Since initiate occurs later in the file - when constructing a
dictionary - it trumps initialize and replaces it.

This patch fixes it by changing the dictionary to a list of verbs
instead of a single stem-verb map.