Remove D413 from the google convention
As none of the examples in the
[Google Python Style Guide](https://google.github.io/styleguide/pyguide.html)
show a docstring with an extra blank line at the end (see the
[3.8.1 Docstrings](https://google.github.io/styleguide/pyguide.html#381-docstrings)
section in particular), error D413 should be excluded from the default
"google" convention.

See #429 for a test case with examples from the Google Python Style
Guide; see also #401 for D413 in relation to the "pep257" convention.