Add the ability to skip checking some functions.
There are times when you might not want to check a function's
docstring.  Specifically, we frequently want to ignore
test files.  There is already an exemption that causes test
functions to be considered non-public.  This extends on
that idea and makes it configurable so a user can cause
a test function to be allowed to have no docstring.