Deprecate pydocstyle in favor of ruff (#658)
Over the last couple of years, I have been the sole maintainer of pydocstyle. During the last year, we made a call for maintainers but we're unsuccessful in finding appropriate maintainers. Given the emergence of ruff and it's features, it's hard to justify the continued maintenance of pydocstyle. After a conversation with @Nurdok we both agree it's time to archive the project in favor of ruff.